Move utils to signer.go
